+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- how `ProductQuantity` worked. Now it is possible for it to receive `selectedItem` and `selectedQuantity` from external sources.
+### Added
+- `BaseProductQuantity` that contains the core logic of `ProductQuantity`
+- `ProductSummaryQuantity`
 
 ## [1.0.3] - 2019-07-02
 ### Fixed